move callbacks test to separate file
